Extract HdfsConfig
