+++ presto-hive/src/main/java/io/prestosql/plugin/hive/HdfsEnvironment.java
@@ -45,7 +45,7 @@ public class HdfsEnvironment
     @Inject
     public HdfsEnvironment(
             HdfsConfiguration hdfsConfiguration,
-            HiveConfig config,
+            HdfsConfig config,
             HdfsAuthentication hdfsAuthentication)
     {
         this.hdfsConfiguration = requireNonNull(hdfsConfiguration, "hdfsConfiguration is null");
