+++ presto-hive/src/test/java/io/prestosql/plugin/hive/TestHiveConfig.java
@@ -13,9 +13,7 @@
  */
 package io.prestosql.plugin.hive;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.net.HostAndPort;
 import io.airlift.units.DataSize;
 import io.airlift.units.DataSize.Unit;
 import io.airlift.units.Duration;
@@ -56,14 +54,6 @@ public class TestHiveConfig
                 .setForceLocalScheduling(false)
                 .setMaxConcurrentFileRenames(20)
                 .setRecursiveDirWalkerEnabled(false)
-                .setDfsTimeout(new Duration(60, TimeUnit.SECONDS))
-                .setIpcPingInterval(new Duration(10, TimeUnit.SECONDS))
-                .setDfsConnectTimeout(new Duration(500, TimeUnit.MILLISECONDS))
-                .setDfsKeyProviderCacheTtl(new Duration(30, TimeUnit.MINUTES))
-                .setDfsConnectMaxRetries(5)
-                .setVerifyChecksum(true)
-                .setDomainSocketPath(null)
-                .setResourceConfigFiles("")
                 .setHiveStorageFormat(HiveStorageFormat.ORC)
                 .setHiveCompressionCodec(HiveCompressionCodec.GZIP)
                 .setRespectTableFormat(true)
@@ -92,13 +82,10 @@ public class TestHiveConfig
                 .setSkipDeletionForAlter(false)
                 .setSkipTargetCleanupOnRollback(false)
                 .setBucketExecutionEnabled(true)
-                .setFileSystemMaxCacheSize(1000)
                 .setTableStatisticsEnabled(true)
                 .setOptimizeMismatchedBucketCount(false)
                 .setWritesToNonManagedTablesEnabled(false)
                 .setCreatesOfNonManagedTablesEnabled(true)
-                .setHdfsSocksProxy(null)
-                .setHdfsWireEncryptionEnabled(false)
                 .setPartitionStatisticsSampleSize(100)
                 .setIgnoreCorruptedStatistics(false)
                 .setRecordingPath(null)
@@ -128,14 +115,6 @@ public class TestHiveConfig
                 .put("hive.per-transaction-metastore-cache-maximum-size", "500")
                 .put("hive.metastore.partition-batch-size.min", "1")
                 .put("hive.metastore.partition-batch-size.max", "1000")
-                .put("hive.dfs.ipc-ping-interval", "34s")
-                .put("hive.dfs-timeout", "33s")
-                .put("hive.dfs.connect.timeout", "20s")
-                .put("hive.dfs.key-provider.cache-ttl", "42s")
-                .put("hive.dfs.connect.max-retries", "10")
-                .put("hive.dfs.verify-checksum", "false")
-                .put("hive.dfs.domain-socket-path", "/foo")
-                .put("hive.config.resources", "/foo.xml,/bar.xml")
                 .put("hive.max-initial-splits", "10")
                 .put("hive.max-initial-split-size", "16MB")
                 .put("hive.split-loader-concurrency", "1")
@@ -173,13 +152,10 @@ public class TestHiveConfig
                 .put("hive.skip-target-cleanup-on-rollback", "true")
                 .put("hive.bucket-execution", "false")
                 .put("hive.sorted-writing", "false")
-                .put("hive.fs.cache.max-size", "1010")
                 .put("hive.table-statistics-enabled", "false")
                 .put("hive.optimize-mismatched-bucket-count", "true")
                 .put("hive.non-managed-table-writes-enabled", "true")
                 .put("hive.non-managed-table-creates-enabled", "false")
-                .put("hive.hdfs.socks-proxy", "localhost:4567")
-                .put("hive.hdfs.wire-encryption.enabled", "true")
                 .put("hive.partition-statistics-sample-size", "1234")
                 .put("hive.ignore-corrupted-statistics", "true")
                 .put("hive.metastore-recording-path", "/foo/bar")
@@ -215,13 +191,6 @@ public class TestHiveConfig
                 .setForceLocalScheduling(true)
                 .setMaxConcurrentFileRenames(100)
                 .setRecursiveDirWalkerEnabled(true)
-                .setIpcPingInterval(new Duration(34, TimeUnit.SECONDS))
-                .setDfsTimeout(new Duration(33, TimeUnit.SECONDS))
-                .setDfsConnectTimeout(new Duration(20, TimeUnit.SECONDS))
-                .setDfsKeyProviderCacheTtl(new Duration(42, TimeUnit.SECONDS))
-                .setDfsConnectMaxRetries(10)
-                .setVerifyChecksum(false)
-                .setResourceConfigFiles(ImmutableList.of("/foo.xml", "/bar.xml"))
                 .setHiveStorageFormat(HiveStorageFormat.SEQUENCEFILE)
                 .setHiveCompressionCodec(HiveCompressionCodec.NONE)
                 .setRespectTableFormat(false)
@@ -230,7 +199,6 @@ public class TestHiveConfig
                 .setMaxPartitionsPerWriter(222)
                 .setMaxOpenSortFiles(333)
                 .setWriteValidationThreads(11)
-                .setDomainSocketPath("/foo")
                 .setTextMaxLineLength(new DataSize(13, Unit.MEGABYTE))
                 .setUseParquetColumnNames(true)
                 .setUseOrcColumnNames(true)
@@ -251,13 +219,10 @@ public class TestHiveConfig
                 .setSkipTargetCleanupOnRollback(true)
                 .setBucketExecutionEnabled(false)
                 .setSortedWritingEnabled(false)
-                .setFileSystemMaxCacheSize(1010)
                 .setTableStatisticsEnabled(false)
                 .setOptimizeMismatchedBucketCount(true)
                 .setWritesToNonManagedTablesEnabled(true)
                 .setCreatesOfNonManagedTablesEnabled(false)
-                .setHdfsSocksProxy(HostAndPort.fromParts("localhost", 4567))
-                .setHdfsWireEncryptionEnabled(true)
                 .setPartitionStatisticsSampleSize(1234)
                 .setIgnoreCorruptedStatistics(true)
                 .setRecordingPath("/foo/bar")
